requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
bcrypt
pandas
pytest
pytest-cov
pyjwt
faker
selenium
Pyarrow